feat: secure user routes, add role-based auth, and build circles/mome…

+++ backend/src/server.ts
@@ -23,7 +23,6 @@ const limiter = rateLimit({
   statusCode: 429,
 });
 
-app.use(limiter);
 app.use(express.json());
 
 // function to geneate a 6 digit verifcation code between 100000 to 999999
@@ -57,6 +56,36 @@ interface UserProp {
   application_id: ApplicationProp;
   otp_attempts: number;
   locked: boolean;
+  role: string;
+}
+
+interface CircleProp {
+  owner_id: string;
+  circle_name: string;
+  circle_image: string;
+  created_at: Date;
+  updated_at: Date;
+}
+
+interface Point {
+  latitude: number;
+  longitude: number;
+}
+
+interface MomentProp {
+  creator_id: string;
+  circle_id: CircleProp;
+  image: string;
+  moments_name: string;
+  created_at: Date;
+  update_at: Date;
+  moment_starts: Date;
+  moment_ends: Date;
+  description: string;
+  location: Point;
+  cap_attendance: number;
+  close_moment: boolean;
+  visibility_type: "nearby" | "circle" | "people";
 }
 
 declare global {
@@ -83,6 +112,24 @@ const authenticateToken = (req: Request, res: Response, next: NextFunction) => {
     req.user = user as UserProp;
     next();
   });
+};
+
+// Defining user role
+const userRole = (role: string) => {
+  return async function (req: Request, res: Response, next: NextFunction) {
+    try {
+      const userRole = req.user?.role;
+      if (role !== userRole) {
+        return res
+          .status(403)
+          .send({ message: "Forbidden. Does not have correct role." });
+      }
+
+      return next();
+    } catch (error) {
+      next(error);
+    }
+  };
 };
 
 //Create Appplication
@@ -137,6 +184,7 @@ app.post(
 app.get(
   "/applications",
   authenticateToken,
+  userRole("admin"),
   async (req: Request, res: Response, next: NextFunction) => {
     try {
       const applications = pool.query("SELECT * FROM applications");
@@ -149,81 +197,80 @@ app.get(
 );
 
 //OTP status merger
-const otpMerge = async (
-  status: string,
-  req: Request<ApplicationProp>,
-  res: Response,
-  next: NextFunction,
-) => {
-  if (!req.body) return res.status(400).send("Request body cannot be empty");
+const otpMerge = (status: string) => {
+  return async function (req: Request, res: Response, next: NextFunction) {
+    if (!req.body) return res.status(400).send("Request body cannot be empty");
 
-  if (status !== "accepted" && status !== "rejected") {
-    return res
-      .status(400)
-      .send(
-        "The server cannot process the request because the client sent invalid syntax or data ",
-      );
-  }
+    if (status !== "accepted" && status !== "rejected") {
+      return res
+        .status(400)
+        .send(
+          "The server cannot process the request because the client sent invalid syntax or data ",
+        );
+    }
 
-  try {
-    const otp = generateOTP();
-    if (!otp) return res.status(401).send("Unauthorized OTP.");
+    try {
+      const otp = generateOTP();
+      if (!otp) return res.status(401).send("Unauthorized OTP.");
 
-    const expiryAt = Date.now() + 300000;
-    const expiryTimeStamp = new Date(expiryAt).toISOString();
+      const expiryAt = Date.now() + 300000;
+      const expiryTimeStamp = new Date(expiryAt).toISOString();
 
-    if (status === "accepted") {
-      const accepted = await pool.query(
-        `INSERT INTO users (application_id, first_name, last_name, email, phone_number, username, otp_code, otp_expiry) 
+      if (status === "accepted") {
+        const accepted = await pool.query(
+          `INSERT INTO users (application_id, first_name, last_name, email, phone_number, username, otp_code, otp_expiry) 
         SELECT id, first_name, last_name, email, phone_number, split_part(email, '@', 1), $2, $3 
         FROM applications 
         WHERE id = $1 
         RETURNING *;`,
-        [req.params.id, otp, expiryTimeStamp],
-      );
+          [req.params.id, otp, expiryTimeStamp],
+        );
 
-      const acceptedUser: UserProp = accepted.rows[0];
-      console.log(acceptedUser);
+        const acceptedUser: UserProp = accepted.rows[0];
+        console.log(acceptedUser);
 
-      await sendSMS({
-        otp_code: otp.toString(),
-        phone_number: acceptedUser.phone_number,
-        status,
-      });
-      await sendEmail({
-        otp_code: otp.toString(),
-        email: acceptedUser.email,
-        status,
-      });
+        await sendSMS({
+          otp_code: otp.toString(),
+          phone_number: acceptedUser.phone_number,
+          status,
+        });
+        await sendEmail({
+          otp_code: otp.toString(),
+          email: acceptedUser.email,
+          status,
+        });
 
-      return res.status(200).send({
-        success: true,
-        data: acceptedUser,
-      });
+        return res.status(200).send({
+          success: true,
+          data: acceptedUser,
+        });
+      }
+
+      if (status === "rejected") {
+        const rejected = await pool.query(
+          `UPDATE users SET status = 'pending' WHERE id = $1 RETURNING *`,
+          [req.params.id],
+        );
+
+        const rejectedUser: UserProp = rejected.rows[0];
+
+        await sendSMS({ phone_number: rejectedUser.phone_number, status });
+        await sendEmail({ email: rejectedUser.email, status });
+
+        return res.status(200).send("Application rejected");
+      }
+    } catch (error) {
+      next(error);
     }
-
-    if (status === "rejected") {
-      const rejected = await pool.query(
-        `UPDATE users SET status = 'pending' WHERE id = $1 RETURNING *`,
-        [req.params.id],
-      );
-
-      const rejectedUser: UserProp = rejected.rows[0];
-
-      await sendSMS({ phone_number: rejectedUser.phone_number, status });
-      await sendEmail({ email: rejectedUser.email, status });
-
-      return res.status(200).send("Application rejected");
-    }
-  } catch (error) {
-    next(error);
-  }
+  };
 };
 
 //Update status applications based on id
 app.put(
   "/applications/:id",
-  async (req: Request<ApplicationProp>, res: Response, next: NextFunction) => {
+  authenticateToken,
+  userRole("admin"),
+  async (req: Request, res: Response, next: NextFunction) => {
     if (!req.body) {
       return res.status(400).send("Request body cannot be empty");
     }
@@ -246,7 +293,7 @@ app.put(
       const updateStatus: ApplicationProp = revealStatus.rows[0];
 
       if (updateStatus) {
-        otpMerge(updateStatus.status, req, res, next);
+        otpMerge(updateStatus.status);
         return;
       }
 
@@ -265,6 +312,11 @@ app.get(
   authenticateToken,
   async (req: Request, res: Response, next: NextFunction) => {
     try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(403)
+          .send("Unauthorized: You can only access your own account");
+      }
       const getUser = await pool.query("SELECT * FROM users WHERE id = $1", [
         req.params.id,
       ]);
@@ -297,9 +349,11 @@ app.post(
   "/auth/verify/",
   limiter,
   async (req: Request, res: Response, next: NextFunction) => {
-    if (!req.query) return res.status(400).send("Request body cannot be empty");
+    if (!req.query)
+      return res.status(400).send("Request query cannot be empty");
 
     const { id } = req.query;
+
     const { otp_code } = req.body;
 
     try {
@@ -310,12 +364,12 @@ app.post(
 
       const lockedUser: UserProp = isUserLocked.rows[0];
 
-      if (lockedUser.locked === true) {
+      if (lockedUser.locked) {
         return res.status(429).send("User Locked!");
       }
 
       const authUser = await pool.query(
-        `SELECT email, otp_code, otp_expiry, username, id, created_at FROM users WHERE id = $1 AND otp_code = $2 AND otp_expiry > NOW()`,
+        `SELECT email, otp_code, otp_expiry, username, id, created_at, role FROM users WHERE id = $1 AND otp_code = $2 AND otp_expiry > NOW()`,
         [id, otp_code],
       );
 
@@ -327,6 +381,7 @@ app.post(
             username: UserAuthorize.username,
             id: UserAuthorize.id,
             created_at: UserAuthorize.created_at,
+            role: UserAuthorize.role,
           },
           `${process.env.JWT_SECRET_KEY}`,
           {
@@ -335,7 +390,7 @@ app.post(
         );
 
         await pool.query(
-          `UPDATE users SET otp_code = null, otp_expiry = null, otp_attempts = 0 WHERE id = $1;`,
+          `UPDATE users SET otp_code = null, otp_expiry = null, otp_attempts = 0 WHERE id = $1 RETURNING *`,
           [id],
         );
 
@@ -380,7 +435,8 @@ app.post(
 //resend OTP request if expired or didn't recieve
 app.put(
   "/auth/resend/:id",
-  async (req: Request<UserProp>, res: Response, next: NextFunction) => {
+  limiter,
+  async (req: Request, res: Response, next: NextFunction) => {
     const updatedGeneratedOTP = generateOTP();
     const updatedExpiryAt = Date.now() + 300000;
     const updatedExpiryTimeStamp = new Date(updatedExpiryAt).toISOString();
@@ -438,16 +494,37 @@ app.put(
   async (req: Request, res: Response, next: NextFunction) => {
     if (!req.body) return res.status(400).send("Request body cannot be empty");
 
-    const userKeys = Object.keys(req.body)
+    const allowed = [
+      "first_name",
+      "last_name",
+      "phone_number",
+      "email",
+      "description",
+      "username",
+    ];
+
+    const filterKeys = Object.keys(req.body).filter((key) => {
+      return allowed.includes(key);
+    });
+
+    const values = filterKeys.map((key) => req.body[key]);
+
+    const userKeys = filterKeys
       .map((key, index) => {
         return `${key} = $${index + 2}`;
       })
       .join(", ");
 
     try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(403)
+          .send("Unauthorized: You can only update your own account");
+      }
+
       const updateUserInformation = await pool.query(
         `UPDATE users SET ${userKeys} WHERE id = $1 RETURNING *`,
-        [req.params.id, ...Object.values(req.body)],
+        [req.params.id, ...values],
       );
 
       const updatedUser: UserProp = updateUserInformation.rows[0];
@@ -474,6 +551,12 @@ app.delete(
   authenticateToken,
   async (req: Request, res: Response, next: NextFunction) => {
     try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(403)
+          .send("Unauthorized: You can only delete your own account");
+      }
+
       const userDeletion = await pool.query(
         `DELETE FROM users WHERE id = $1 RETURNING *`,
         [req.params.id],
@@ -493,6 +576,212 @@ app.delete(
         failed: true,
         message: `User Doesn't Exist: ${req.params.id}`,
       });
+    } catch (error) {
+      next(error);
+    }
+  },
+);
+
+//Creating a Circle based on User id
+app.post(
+  "/circles/:id",
+  authenticateToken,
+  async (req: Request, res: Response, next: NextFunction) => {
+    if (!req.body) return res.status(400).send("Request body cannot be empty.");
+
+    const { circle_name } = req.body;
+
+    if (!circle_name) {
+      return res.status(400).send(`${circle_name} is required.`);
+    }
+
+    try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(400)
+          .send("Unauthorized: You can only access your own account");
+      }
+      const insertCircle = await pool.query(
+        "INSERT INTO circles (owner_id, circle_name) VALUES ($1, $2) RETURNING *",
+        [req.params.id, circle_name],
+      );
+
+      const createCircle: CircleProp = await insertCircle.rows[0];
+
+      if (createCircle) {
+        return res.status(200).send({
+          success: true,
+          data: createCircle,
+        });
+      }
+
+      return res.status(400).send("Unable to Create Circle");
+    } catch (error) {
+      next(error);
+    }
+  },
+);
+
+//Retrieving all Circles based on User id
+app.get(
+  "/circles/:id",
+  authenticateToken,
+  async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(400)
+          .send("Unauthorized: You can only access your own account");
+      }
+
+      const getCircleById = await pool.query(
+        `SELECT * FROM circles WHERE owner_id = $1`,
+        [req.params.id],
+      );
+
+      const retrieveCircle: CircleProp[] = getCircleById.rows;
+
+      if (retrieveCircle) {
+        return res.status(200).send({
+          success: true,
+          data: retrieveCircle,
+        });
+      }
+
+      return res.status(400).send("Unable to retrieve circles");
+    } catch (error) {
+      next(error);
+    }
+  },
+);
+
+//Updating circles based on id and cirle
+app.put(
+  "/circles/:id/:circle_id",
+  authenticateToken,
+  async (req: Request, res: Response, next: NextFunction) => {
+    if (!req.body) return res.status(400).send("Request body cannot be empty");
+
+    const allowed = ["circle_name", "circle_image"];
+
+    const filterKeys = Object.keys(req.body).filter((key) => {
+      // just filters the key for you
+      return allowed.includes(key);
+    });
+
+    const values = filterKeys.map((value) => req.body[value]); //takes in the values of the filtered key
+
+    const circleKeys = filterKeys
+      .map((circle, index) => {
+        return `${circle} = $${index + 3}`;
+      })
+      .join(", ");
+
+    try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(403)
+          .send("Unauthorized: You can only delete your own circle");
+      }
+      const updateCircle = await pool.query(
+        `UPDATE circles SET ${circleKeys} WHERE owner_id = $1 AND id = $2 RETURNING *`,
+        [req.params.id, req.params.circle_id, ...values],
+      );
+
+      const circleUpdated: CircleProp = updateCircle.rows[0];
+
+      if (circleUpdated) {
+        return res.status(200).send({
+          success: true,
+          data: {
+            circle_name: circleUpdated.circle_name,
+            circle_image: circleUpdated.circle_image,
+          },
+        });
+      }
+
+      return res.status(404).send("Desired Circle doesn't exist");
+    } catch (error) {
+      next(error);
+    }
+  },
+);
+
+//Deleting circles based on id and cirle
+app.delete(
+  "/circles/:id/:circle_id",
+  authenticateToken,
+  async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(403)
+          .send("Unauthorized: You can only delete your own circle");
+      }
+
+      const deleteCircleByUser = await pool.query(
+        `DELETE FROM circles WHERE owner_id = $1 AND id = $2 RETURNING *`,
+        [req.params.id, req.params.circle_id],
+      );
+
+      const deleteCircle: CircleProp = deleteCircleByUser.rows[0];
+
+      if (deleteCircle) {
+        return res.status(200).send({
+          success: true,
+          message: `Circle named ${deleteCircle.circle_name} successfully deleted`,
+        });
+      }
+
+      return res.status(400).send("Unable to handle request");
+    } catch (error) {
+      next(error);
+    }
+  },
+);
+
+//Creating a Moment based on User id
+app.post(
+  "/moments/:id",
+  authenticateToken,
+  async (req: Request, res: Response, next: NextFunction) => {
+    if (!req.body) return res.status(400).send("Request body cannot be empty.");
+
+    const { moments_name, location, moment_starts, visibility_type } = req.body;
+
+    if (!moments_name || !location || !moment_starts || !visibility_type) {
+      const missing = [
+        "moments_name",
+        "location",
+        "moment_starts",
+        "visibility_type",
+      ].find((key) => !req.body[key]);
+
+      return res.status(400).send(`${missing} is required.`);
+    }
+
+    try {
+      if (req.user?.id !== req.params.id) {
+        return res
+          .status(400)
+          .send("Unauthorized: You can only access your own account");
+      }
+
+      const createMomentsById = await pool.query(
+        "INSERT INTO moments (creator_id, moments_name, location, moment_starts, visibility_type) VALUES ($1, $2, $3, $4, $5) RETURNING *",
+        [req.params.id, moments_name, location, moment_starts, visibility_type],
+      );
+
+      const createMoment: MomentProp = createMomentsById.rows[0];
+
+      if (createMoment) {
+        return res.status(200).send({
+          success: true,
+          createMoment: createMoment,
+        });
+      }
+
+      return res.status(400).send("Unable to Create Moment");
     } catch (error) {
       next(error);
     }
